+++ agents/relayer/src/main.rs
@@ -14,14 +14,29 @@ use crate::{relayer::Relayer, settings::RelayerSettings as Settings};
 use color_eyre::Result;
 use nomad_base::NomadAgent;
 
+use tracing::info_span;
+use tracing_subscriber::prelude::*;
+
 #[tokio::main(flavor = "current_thread")]
 async fn main() -> Result<()> {
     color_eyre::install()?;
-    let settings = Settings::new()?;
 
+    // sets the subscriber for this scope only
+    let _bootup_guard = tracing_subscriber::FmtSubscriber::builder()
+        .json()
+        .with_level(true)
+        .set_default();
+
+    let span = info_span!("RelayerBootup");
+    let _span = span.enter();
+
+    let settings = Settings::new().await?;
     let agent = Relayer::from_settings(settings).await?;
 
-    agent.start_tracing(agent.metrics().span_duration())?;
+    drop(_span);
+    drop(span);
+
+    let _tracing_guard = agent.start_tracing(agent.metrics().span_duration());
 
     let _ = agent.metrics().run_http_server();