+++ nomad-base/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+- add home and remote labels to contract sync metrics for event differentiation
+- add `CONFIG_URL` check to `decl_settings` to optionally fetch config from a remote url
 - bug: add checks for empty replica name arrays in `NomadAgent::run_many` and
   `NomadAgent::run_all`
 - add `previously_attempted` to the DB schema